@@ -4,42 +4,42 @@
     "version": "1.0.0",
     "author": "Derek Kershner",
     "dependencies": {
-        "bootstrap": "^4.3.1",
-        "bulma": "^0.7.0",
-        "gatsby": "^2.13.31",
-        "gatsby-image": "^2.0.23",
-        "gatsby-plugin-google-analytics": "^2.1.7",
-        "gatsby-plugin-netlify": "^2.0.6",
-        "gatsby-plugin-netlify-cms": "^4.1.6",
-        "gatsby-plugin-purgecss": "^4.0.0",
-        "gatsby-plugin-react-helmet": "^3.0.4",
-        "gatsby-plugin-sass": "^2.0.7",
-        "gatsby-plugin-sharp": "^2.2.9",
-        "gatsby-plugin-sitemap": "^2.2.6",
-        "gatsby-plugin-typescript": "^2.1.2",
-        "gatsby-remark-copy-linked-files": "^2.0.7",
-        "gatsby-remark-images": "^3.1.6",
-        "gatsby-remark-relative-images": "^0.2.1",
-        "gatsby-source-filesystem": "^2.0.26",
-        "gatsby-transformer-remark": "^2.6.9",
-        "gatsby-transformer-sharp": "^2.1.9",
+        "bootstrap": "^4.4.1",
+        "bulma": "^0.8.0",
+        "gatsby": "^2.19.12",
+        "gatsby-image": "^2.2.40",
+        "gatsby-plugin-google-analytics": "^2.1.35",
+        "gatsby-plugin-netlify": "^2.1.32",
+        "gatsby-plugin-netlify-cms": "^4.1.38",
+        "gatsby-plugin-purgecss": "^4.0.1",
+        "gatsby-plugin-react-helmet": "^3.1.22",
+        "gatsby-plugin-sass": "^2.1.28",
+        "gatsby-plugin-sharp": "^2.4.5",
+        "gatsby-plugin-sitemap": "^2.2.27",
+        "gatsby-plugin-typescript": "^2.1.27",
+        "gatsby-remark-copy-linked-files": "^2.1.37",
+        "gatsby-remark-images": "^3.1.44",
+        "gatsby-remark-relative-images": "^0.2.3",
+        "gatsby-source-filesystem": "^2.1.48",
+        "gatsby-transformer-remark": "^2.6.50",
+        "gatsby-transformer-sharp": "^2.3.14",
         "lodash": "^4.17.15",
-        "lodash-webpack-plugin": "^0.11.4",
-        "netlify-cms-app": "^2.9.6",
-        "netlify-cms-media-library-cloudinary": "^1.3.2",
-        "netlify-cms-media-library-uploadcare": "^0.5.2",
-        "node-sass": "^4.11.0",
-        "parcel-bundler": "^1.9.4",
-        "react": "^16.8.4",
+        "lodash-webpack-plugin": "^0.11.5",
+        "netlify-cms-app": "^2.11.14",
+        "netlify-cms-media-library-cloudinary": "^1.3.4",
+        "netlify-cms-media-library-uploadcare": "^0.5.5",
+        "node-sass": "^4.13.1",
+        "parcel-bundler": "^1.12.4",
+        "react": "^16.12.0",
         "react-arrow": "^2.0.4",
-        "react-bootstrap": "^1.0.0-beta.6",
-        "react-dom": "^16.8.4",
-        "react-helmet": "^5.2.0",
-        "react-icons": "^3.7.0",
+        "react-bootstrap": "^1.0.0-beta.16",
+        "react-dom": "^16.12.0",
+        "react-helmet": "^5.2.1",
+        "react-icons": "^3.9.0",
         "react-lazy-hero": "^2.0.0",
-        "react-scroll": "^1.7.12",
-        "styled-components": "^4.2.0",
-        "uuid": "^3.2.1"
+        "react-scroll": "^1.7.16",
+        "styled-components": "^5.0.1",
+        "uuid": "^3.4.0"
     },
     "keywords": [
         "gatsby"
@@ -58,37 +58,38 @@
         "serve": "gatsby serve",
         "format": "prettier --trailing-comma es5 --no-semi --single-quote --write \"{gatsby-*.js,src/**/*.js}\"",
         "test": "jest --coverage",
-        "lint": "eslint ./src --ext .tsx --ext .js"
+        "lint": "eslint ./src --ext .tsx --ext .js",
+        "type-check": "tsc --noEmit"
     },
     "devDependencies": {
-        "@testing-library/jest-dom": "^4.0.0",
-        "@testing-library/react": "^9.1.1",
-        "@types/bootstrap": "^4.3.0",
-        "@types/jest": "^24.0.18",
-        "@types/node": "^12.7.2",
-        "@types/react": "^16.9.2",
-        "@types/react-bootstrap": "^0.32.17",
-        "@types/react-dom": "^16.8.5",
-        "@types/react-helmet": "^5.0.8",
+        "@testing-library/jest-dom": "^5.1.1",
+        "@testing-library/react": "^9.4.0",
+        "@types/bootstrap": "^4.3.1",
+        "@types/jest": "^25.1.2",
+        "@types/node": "^13.7.0",
+        "@types/react": "^16.9.19",
+        "@types/react-bootstrap": "^0.32.20",
+        "@types/react-dom": "^16.9.5",
+        "@types/react-helmet": "^5.0.15",
         "@types/react-router-bootstrap": "^0.24.5",
-        "@types/react-router-dom": "^4.3.3",
+        "@types/react-router-dom": "^5.1.3",
         "@types/react-scroll": "^1.5.4",
-        "@typescript-eslint/eslint-plugin": "^1.11.0",
-        "@typescript-eslint/parser": "^1.11.0",
-        "babel-jest": "^24.9.0",
-        "babel-preset-gatsby": "^0.2.8",
-        "eslint": "^6.1.0",
-        "eslint-config-prettier": "^6.0.0",
-        "eslint-plugin-prettier": "^3.1.0",
-        "eslint-plugin-react": "^7.14.2",
+        "@typescript-eslint/eslint-plugin": "^2.19.0",
+        "@typescript-eslint/parser": "^2.19.0",
+        "babel-jest": "^25.1.0",
+        "babel-preset-gatsby": "^0.2.28",
+        "eslint": "^6.8.0",
+        "eslint-config-prettier": "^6.10.0",
+        "eslint-plugin-prettier": "^3.1.2",
+        "eslint-plugin-react": "^7.18.3",
         "esm": "^3.2.25",
-        "http-proxy-middleware": "^0.19.1",
+        "http-proxy-middleware": "^0.20.0",
         "identity-obj-proxy": "^3.0.0",
-        "jest": "^24.9.0",
-        "netlify-lambda": "^1.4.3",
+        "jest": "^25.1.0",
+        "netlify-lambda": "^1.6.3",
         "npm-run-all": "^4.1.5",
-        "prettier": "^1.15.3",
-        "ts-jest": "^24.0.2",
-        "typescript": "^3.5.3"
+        "prettier": "^1.19.1",
+        "ts-jest": "^25.2.0",
+        "typescript": "^3.7.5"
     }
 }
