@@ -1,12 +1,12 @@
-import React from 'react';
+import React, { ReactElement } from 'react';
 import { Jumbotron } from 'react-bootstrap';
 
 import PortfolioHeader from './portfolio/PortfolioHeader';
 import { projects } from '../data/Projects';
 
-const PortfolioMenu = (): JSX.Element => {
+const PortfolioMenu = (): ReactElement => {
     return (
-        <React.Fragment>
+        <>
             <Jumbotron id="portfolio" data-testid="portfolio-menu" fluid className="bg-primary text-white">
                 <h3 className="display-4 text-center">
                     <a className="text-light" href="/portfolio">
@@ -15,12 +15,10 @@ const PortfolioMenu = (): JSX.Element => {
                 </h3>
             </Jumbotron>
 
-            {projects.map(
-                (project, index): JSX.Element => (
-                    <PortfolioHeader key={index} project={project} menu={true} />
-                )
-            )}
-        </React.Fragment>
+            {projects.map((project, index) => (
+                <PortfolioHeader key={index} project={project} menu={true} />
+            ))}
+        </>
     );
 };
 
